# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskManager DOC API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/tasks": {
            "get": {
                "tags": [
                    "Админские функции (Данные Заданий)"
                ],
                "summary": "Список Заданий всех пользователей",
                "operationId": "a1e82472d5018817ab186e53c08f538b",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Id user",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Название Задания",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Описание Задания",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Статус задания,\n *         1-Создано\n *         2-Исполнено\n *         3-Удалено",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "created_from",
                        "in": "query",
                        "description": "Дата создания с",
                        "required": false,
                        "schema": {
                            "type": "date-time"
                        }
                    },
                    {
                        "name": "created_to",
                        "in": "query",
                        "description": "Дата создания до",
                        "required": false,
                        "schema": {
                            "type": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Переустановить что-то"
                                                    },
                                                    "content": {
                                                        "type": "text",
                                                        "example": "Способ переустановки"
                                                    },
                                                    "status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2010-10-22"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Админские функции (Данные Заданий)"
                ],
                "summary": "Создание 'Задания'",
                "operationId": "2c22522bbdc8f90794896a2d030dc83b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Переустановить MacOs"
                                            },
                                            "content": {
                                                "type": "text",
                                                "example": "Действия по заданию"
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Переустановить MacOs"
                                        },
                                        "content": {
                                            "type": "text",
                                            "example": "Действия по заданию"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2010-10-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unautoruzed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/tasks/{task}": {
            "get": {
                "tags": [
                    "Админские функции (Данные Заданий)"
                ],
                "summary": "Просмотр Задания пользователя",
                "operationId": "74559716cf71469c90898487f292e5d6",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "id task",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Переустановить MacOs"
                                        },
                                        "content": {
                                            "type": "text",
                                            "example": "Действия по заданию"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2010-10-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Админские функции (Данные Заданий)"
                ],
                "summary": "Удаление 'Задание'",
                "operationId": "5cd53b6d0b643d0b9a6e51377a08bd3d",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "id task",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": " Task Deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Админские функции (Данные Заданий)"
                ],
                "summary": "Обновление данных Задания",
                "operationId": "cb21fc827923c85d4394a8e91fd37923",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "id task",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Переустановить MacOs"
                                            },
                                            "content": {
                                                "type": "text",
                                                "example": "Действия по заданию"
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Переустановить MacOs"
                                                    },
                                                    "content": {
                                                        "type": "text",
                                                        "example": "Действия по заданию"
                                                    },
                                                    "status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2010-10-22"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unautoruzed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}